convertnumber

+++ convert_number.c
@@ -16,7 +16,7 @@ int print_hex(va_list ap, params_t *params)
 	if (params->l_modifier)
 		l = (unsigned long)va_arg(ap, unsigned long);
 	else if (params->h_modifier)
-		l = (unsigned long)va_arg(ap, unsigned int);
+		l = (unsigned short int)va_arg(ap, unsigned int);
 	else
 		l = (unsigned int)va_arg(ap, unsigned int);
 
@@ -38,16 +38,16 @@ int print_hex(va_list ap, params_t *params)
  */
 int print_HEX(va_list ap, params_t *params)
 {
-unsigned long l;
+	unsigned long l;
 	int c = 0;
 	char *str;
 
 	if (params->l_modifier)
-	l = (unsigned long)va_arg(ap, unsigned long);
+		l = (unsigned long)va_arg(ap, unsigned long);
 	else if (params->h_modifier)
-	l = (unsigned long)va_arg(ap, unsigned int);
+		l = (unsigned short int)va_arg(ap, unsigned int);
 	else
-	l = (unsigned int)va_arg(ap, unsigned int);
+		l = (unsigned int)va_arg(ap, unsigned int);
 
 	str = convert(l, 16, CONVERT_UNSIGNED, params);
 	if (params->hashtag_flag && l)
@@ -67,16 +67,16 @@ unsigned long l;
  */
 int print_binary(va_list ap, params_t *params)
 {
-unsigned int n = va_arg(ap, unsigned int);
-char *str = convert(n, 2, CONVERT_UNSIGNED, params);
-int c = 0;
+	unsigned int n = va_arg(ap, unsigned int);
+	char *str = convert(n, 2, CONVERT_UNSIGNED, params);
+	int c = 0;
 
-if (params->hashtag_flag && n)
-{
-*--str = '0';
-}
-params->unsign = 1;
-return (c += print_number(str, params));
+	if (params->hashtag_flag && n)
+	{
+		*--str = '0';
+	}
+	params->unsign = 1;
+	return (c += print_number(str, params));
 }
 
 /**
@@ -95,7 +95,7 @@ int print_octal(va_list ap, params_t *params)
 	if (params->l_modifier)
 		l = (unsigned long)va_arg(ap, unsigned long);
 	else if (params->h_modifier)
-		l = (unsigned short int)va_arg(ap, unsigned long);
+		l = (unsigned short int)va_arg(ap, unsigned int);
 	else
 		l = (unsigned int)va_arg(ap, unsigned int);
 	str = convert(l, 8, CONVERT_UNSIGNED, params);
